ssh-spawner: resolve a printf format warning
On Solaris/ILP32 pid_t is typedef'ed to long (but is int for LP64).
So, this commit adds a cast to avoid a printf format warning when
printing a pid_t with format "%d'.

+++ other/ssh-spawner/gasnet_bootstrap_ssh.c
@@ -572,7 +572,7 @@ static void reap_one(pid_t pid, int status)
       }
     }
     if (j == children) {
-      BOOTSTRAP_VERBOSE(("[%d] Reaped unknown pid %d\n", myname, pid));
+      BOOTSTRAP_VERBOSE(("[%d] Reaped unknown pid %d\n", myname, (int)pid));
     }
   }
